[level 2] Title: 더 맵게, Time: 1813.49 ms, Memory: 123 MB -BaekjoonHub

+++ 프로그래머스/2/42626. 더 맵게/더 맵게.java
@@ -2,15 +2,15 @@ import java.util.*;
 class Solution {
     public int solution(int[] scoville, int K) {
         int cnt = 0;
-        PriorityQueue<Long> pq = new PriorityQueue<>();
+        PriorityQueue<Integer> pq = new PriorityQueue<>();
         for(int s : scoville) {
-            pq.add((long)s);
+            pq.add(s);
         }
         while(pq.size() > 1) {
             if(pq.peek() >= K) {
                 break;
             }
-            long result = pq.poll() + pq.poll() * 2;
+            int result = pq.poll() + pq.poll() * 2;
             cnt++;
             pq.add(result);
